+++ nekoyume/Assets/_Scripts/UI/Module/MailReward.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using Libplanet.Types.Assets;
 using Nekoyume.Blockchain;
 using Nekoyume.Game;
@@ -11,18 +12,18 @@ namespace Nekoyume.UI.Module
     {
         public ItemBase ItemBase { get; }
         public FungibleAssetValue FavFungibleAssetValue { get; }
-        public long Count { get; }
+        public BigInteger Count { get; }
 
         public bool IsPurchased { get; }
 
-        public MailReward(ItemBase itemBase, long count, bool isPurchased = false)
+        public MailReward(ItemBase itemBase, BigInteger count, bool isPurchased = false)
         {
             ItemBase = itemBase;
             Count = count;
             IsPurchased = isPurchased;
         }
 
-        public MailReward(FungibleAssetValue fav, long count, bool isPurchased = false)
+        public MailReward(FungibleAssetValue fav, BigInteger count, bool isPurchased = false)
         {
             FavFungibleAssetValue = fav;
             Count = count;
